refactor: переименование файлов компонентов
Файлы для Preact компонетов должны именоваться с заглавной буквы для
лучшей ясности. Этот коммит исправляет эту проблему.
