+++ src/components/box/FollowText.tsx
