+++ src/components/vk/Hint.tsx
