+++ src/components/vk/ProgressIndicator.tsx
